Remove unneeded apt_source dependency (#4)
The apt_source dependency was used until Debian 8 (Jessie) for dotdeb
repository. However, dotdeb is not updated anymore and it does not
support Debian 9 (Stretch) or later. Therefore, dotdeb is not needed
anymore.

The support for Jessie has ceased in general and this role does not need
to be compatible with it.

The Ubuntu config is not actively used either and can be removed.

Overall, apt_source is not needed anymore in the role.